refactor: 컨트롤러 및 DTO 패키지 구조 리팩토링 및 RestController로 변경
