+++ src/main/java/roomescape/controller/LoginController.java
@@ -4,18 +4,18 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Member;
 import roomescape.dto.request.LoginRequest;
 import roomescape.dto.response.MemberResponse;
 import roomescape.service.AuthService;
 import roomescape.service.MemberService;
 
-@Controller
+@RestController
 @RequestMapping("/login")
 public class LoginController {
 
@@ -25,12 +25,6 @@ public class LoginController {
     public LoginController(MemberService memberService, AuthService authService) {
         this.memberService = memberService;
         this.authService = authService;
-    }
-
-    // TODO: 페이지 컨트롤러 위치 고려해보기
-    @GetMapping
-    public String getLoginPage() {
-        return "login";
     }
 
     @PostMapping
