+++ src/main/java/roomescape/controller/page/AdminPageController.java
@@ -1,0 +1,30 @@
+package roomescape.controller.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminPageController {
+
+    @GetMapping
+    public String adminMainPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "admin/reservation-new";
+    }
+
+    @GetMapping("/time")
+    public String reservationTimePage() {
+        return "admin/time";
+    }
+
+    @GetMapping("/theme")
+    public String themePage() {
+        return "admin/theme";
+    }
+}
