+++ src/main/java/roomescape/controller/page/MainPageController.java
@@ -1,10 +1,15 @@
-package roomescape.controller;
+package roomescape.controller.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class MainPageController {
+
+    @GetMapping("login")
+    public String getLoginPage() {
+        return "login";
+    }
 
     @GetMapping("/reservation")
     public String mainPage() {
