+++ src/main/java/roomescape/dto/request/AdminCreateReservationRequest.java
@@ -2,8 +2,7 @@ package roomescape.dto.request;
 
 import java.time.LocalDate;
 
-// TODO: 이름 고려해보기
-public record CreateReservationByAdminRequest(
+public record AdminCreateReservationRequest(
         Long memberId,
         Long timeId,
         Long themeId,
